requirements.txt
Django==2.2.2
djangorestframework==3.9.4
psycopg2-binary==2.8.3